+++ src/assets/APDevLogo.tsx
@@ -1,5 +1,3 @@
-const APDevLogoOpenSans = "";
-
 export const APDevLogo = () => {
   return (
     <div className="logo">
