+++ src/components/Footer.tsx
@@ -1,8 +1,7 @@
 import GitHubIcon from '@mui/icons-material/GitHub';
 import LinkedInIcon from '@mui/icons-material/LinkedIn';
-import FilePresentIcon from '@mui/icons-material/FilePresent';
 import ViewCarouselIcon from '@mui/icons-material/ViewCarousel';
-import resume from "../assets/AndyPrattResume.pdf"; // Cannot find module './resume.pdf'.ts(2307)
+
 
 export const Footer = () => {
 
@@ -20,10 +19,6 @@ export const Footer = () => {
 
         <a className="footer-links" href="https://www.linkedin.com/in/andrewpratt37/" target="_blank" rel="noreferrer noopener" style={{margin: 10}}>
           <LinkedInIcon fontSize='medium' color="inherit" /> LinkedIn |
-        </a>
-
-        <a className="footer-links" href={resume} target="_blank" rel="noreferrer noopener" style={{margin: 10}}>
-          <FilePresentIcon fontSize='medium' color="inherit"/>Resume |
         </a>
 
         <a className="footer-links" href={"http://cardswithfriendsgame.com"} target="_blank" rel="noreferrer noopener" style={{margin: 10}}>
