+++ src/components/OldNavbar.tsx
@@ -1,0 +1,134 @@
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
+import Menu from '@mui/material/Menu';
+import Button from '@mui/material/Button';
+import MenuItem from '@mui/material/MenuItem';
+import MenuIcon from '@mui/icons-material/Menu';
+import { Link } from "react-router-dom"
+import { APDevLogo } from '../assets/APDevLogo';
+
+export const Navbar = () => {
+  const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(null);
+  const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(null);
+
+  const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorElNav(event.currentTarget);
+  };
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+  
+  return (
+    <AppBar style={{backgroundColor: "#292929", fontStyle: "Roboto"}} color="inherit" position="static">
+      <div className="navbar-container">
+        <Toolbar disableGutters>
+
+
+          <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
+
+          <Link to="/">
+            <APDevLogo/>
+          </Link>
+          {/* APP BAR */}
+
+            <IconButton
+              size="large"
+              aria-label="account of current user"
+              aria-controls="menu-appbar"
+              aria-haspopup="true"
+              onClick={handleOpenNavMenu}
+              color="inherit"
+            >
+              <MenuIcon />
+            </IconButton>
+            
+            {/* App Bar Menu Items */}
+            <Menu
+              id="menu-appbar"
+              anchorEl={anchorElNav}
+              anchorOrigin={{
+                vertical: 'bottom',
+                horizontal: 'left',
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: 'top',
+                horizontal: 'left',
+              }}
+              open={Boolean(anchorElNav)}
+              onClose={handleCloseNavMenu}
+              sx={{
+                display: { xs: 'block', md: 'none' },
+              }}
+            >
+              <MenuItem>
+                <Typography textAlign="center">Portfolio</Typography>
+              </MenuItem>
+              <MenuItem>
+                <Typography textAlign="center">Resume</Typography>
+              </MenuItem>
+              <MenuItem>
+                <Typography textAlign="center">About Me</Typography>
+              </MenuItem>
+            </Menu>
+          </Box>
+
+          {/* DESKTOP NAV BAR */}
+
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}
+          >
+
+          <Link to="/">
+            <APDevLogo/>
+          </Link>
+
+          </Typography>
+          <Box className="navbar-button-container" sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
+            <Button sx={{ my: 2, color: 'inherit', display: 'block' }}>
+              Portfolio
+            </Button>
+            <Button sx={{ my: 2, color: 'inherit', display: 'block' }}>
+              Resume
+            </Button>
+            <Button sx={{ my: 2, color: 'inherit', display: 'block' }}>
+              About
+            </Button>
+          </Box>
+
+          <Box sx={{ flexGrow: 0 }}>
+            <Menu
+              sx={{ mt: '45px' }}
+              id="menu-appbar"
+              anchorEl={anchorElUser}
+              anchorOrigin={{
+                vertical: 'top',
+                horizontal: 'right',
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: 'top',
+                horizontal: 'right',
+              }}
+              open={Boolean(anchorElUser)}
+              onClose={handleCloseUserMenu}
+            >
+            </Menu>
+          </Box>
+        </Toolbar>
+      </div>
+    </AppBar>
+  );
+};
