+++ src/styles/App.css
@@ -4,12 +4,45 @@ body {
   color: white;
 }
 
-.logo {
-  background-color: #292929;
-  cursor: pointer;
+.app {
 }
 
-.app {
+/* NAVBAR */
+
+a {
+  text-decoration: none;
+}
+
+.logo {
+  align-self: left;
+}
+
+.desktop-nav {
+  text-align: center;
+  justify-content: center;
+  height: 50px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.desktop-nav-button-container {
+  width: 300px;
+  display: flex;
+  flex-direction: row;
+  align-self: "right";
+  justify-content: space-evenly;
+}
+
+
+.desktop-nav-button {
+  color: white;
+  text-decoration: none;
+  font-family: Roboto, Helvetica, Arial, sans-serif;
+  font-weight: 500;
+  font-size: 0.875rem;
+  line-height: 1.75;
+  letter-spacing: 0.02857em;
 }
 
 .page-container {
@@ -50,7 +83,7 @@ body {
   font-size: 1.75rem;
 }
 
-/* Footer & Footer Links */
+/* Footer */
 
 .footer {
   position: absolute;
